Refactoring of `pathfinder` interface layer.
- Moved `pathfinder_db.rs` module to it's own folder for future
extension with other `pathfinder` related bindings.
- Renamed `visited_pcs.rs` to `replay_class_hash.rs` to match
structure name.
