+++ starknet-replay/src/lib.rs
@@ -40,13 +40,14 @@ use error::RunnerError;
 use runner::replay_block::ReplayBlock;
 
 pub use crate::histogram::export as export_histogram;
+pub use crate::pathfinder_storage::{connect_to_database, get_latest_block_number};
 pub use crate::profiler::replay_statistics::ReplayStatistics;
 pub use crate::profiler::report::write_to_file;
-pub use crate::runner::pathfinder_db::{connect_to_database, get_latest_block_number};
 pub use crate::runner::replay_range::ReplayRange;
 pub use crate::runner::run_replay;
 
 pub mod error;
 pub mod histogram;
+pub mod pathfinder_storage;
 pub mod profiler;
 pub mod runner;
